Docker it
