+++ dependabot_on/Dockerfile
@@ -1,0 +1,6 @@
+FROM python:3.8.0rc1-alpine
+
+COPY requirements.txt .
+RUN pip install -r requirements.txt
+
+COPY . .
