+++ dependabot_on/language_lookup.py
@@ -1,0 +1,71 @@
+# One of bundler, composer, docker, maven, npm_and_yarn,
+# elm, submodules, hex, cargo, gradle, nuget, dep, go_modules,
+# pip, terraform, github_actions
+
+package_managers = {
+        "ApacheConf": "",
+        "API Blueprint": "",
+        "ASP": "",
+        "Assembly": "",
+        "Awk": "",
+        "Batchfile": "",
+        "C": "",
+        "C#": "nuget",
+        "C++": "",
+        "Clojure": "",
+        "CoffeeScript": "npm_and_yarn",
+        "CSS": "",
+        "Cycript": "",
+        "Dockerfile": "docker",
+        "Elixir": "",
+        "Emacs Lisp": "",
+        "FreeMarker": "",
+        "Gherkin": "",
+        "Go": "go_modules",
+        "Groovy": "",
+        "Handlebars": "",
+        "Haskell": "",
+        "HCL": "terraform",
+        "HTML": "",
+        "Inno Setup": "",
+        "Java": "maven",
+        "JavaScript": "npm_and_yarn",
+        "JSONiq": "",
+        "Jsonnet": "",
+        "Jupyter Notebook": "pip",
+        "Lex": "",
+        "Liquid": "",
+        "Logos": "",
+        "Lua": "",
+        "M4": "",
+        "Makefile": "",
+        "Mako": "",
+        "Nginx": "",
+        "Nix": "",
+        "Pascal": "",
+        "Perl": "",
+        "PHP": "composer",
+        "PLpgSQL": "",
+        "PLSQL": "",
+        "PowerShell": "",
+        "Protocol Buffer": "",
+        "Puppet": "",
+        "Python": "pip",
+        "R": "",
+        "Ragel": "",
+        "Roff": "",
+        "Ruby": "bundler",
+        "Scala": "",
+        "Scheme": "",
+        "sed": "",
+        "Shell": "",
+        "Smarty": "",
+        "Tcl": "",
+        "TSQL": "",
+        "TypeScript": "",
+        "VCL": "",
+        "Vim script": "",
+        "Visual Basic": "",
+        "XSLT": "",
+        "Yacc": "",
+    }
