+++ dependabot_on/requirements.txt
@@ -1,0 +1,2 @@
+addict
+requests
